--- a/libraries/Servo/examples/Knob/Knob.ino
+++ b/libraries/Servo/examples/Knob/Knob.ino
@@ -27,5 +27,5 @@ void loop() {
     val = analogRead(potpin); // reads the value of the potentiometer
     val = map(val, 0, getAnalogReadMaximum(), 0, 180); // scale it to use it with the servo (value between 0 and 180)
     myservo.write(val); // sets the servo position according to the scaled value
-    delay(15); // waits for the servo to get there
+    delay(20); // waits for the servo to get there
 }