--- a/libraries/Servo/examples/Sweep/Sweep.ino
+++ b/libraries/Servo/examples/Sweep/Sweep.ino
@@ -13,27 +13,24 @@
 
 #include <Servo.h>
 
-
-#define RESOLUTION 16 // sets PWM resolution to 16 bits (65535 max)
-
-Servo myservo;  // create servo object to control a servo
+Servo myservo; // create servo object to control a servo
 // twelve servo objects can be created on most boards
 
-int pos = 0;    // variable to store the servo position
-
 void setup() {
-  analogWriteResolution(RESOLUTION); //Needed for better accuracy when driving a Servo
-  myservo.attach(9);  // attaches the servo on pin 9 to the servo object
+    analogWriteResolution(16); // sets PWM resolution to 16 bits, for better servo accuracy
+    myservo.attach(9); // attaches the servo on pin 9 to the servo object
 }
 
 void loop() {
-  for (pos = 0; pos <= 180; pos += 1) { // goes from 0 degrees to 180 degrees
-    // in steps of 1 degree
-    myservo.write(pos);              // tell servo to go to position in variable 'pos'
-    delay(15);                       // waits 15ms for the servo to reach the position
-  }
-  for (pos = 180; pos >= 0; pos -= 1) { // goes from 180 degrees to 0 degrees
-    myservo.write(pos);              // tell servo to go to position in variable 'pos'
-    delay(15);                       // waits 15ms for the servo to reach the position
-  }
+    int pos; // variable to store the servo position
+    for (pos = 0; pos <= 180; pos += 1) { // goes from 0 degrees to 180 degrees
+        // in steps of 1 degree
+        myservo.write(pos); // tell servo to go to position in variable 'pos'
+        delay(15); // waits 15ms for the servo to reach the position
+    }
+    
+    for (pos = 180; pos >= 0; pos -= 1) { // goes from 180 degrees to 0 degrees
+        myservo.write(pos); // tell servo to go to position in variable 'pos'
+        delay(15); // waits 15ms for the servo to reach the position
+    }
 }