--- a/libraries/Servo/examples/Sweep/Sweep.ino
+++ b/libraries/Servo/examples/Sweep/Sweep.ino
@@ -26,11 +26,11 @@ void loop() {
     for (pos = 0; pos <= 180; pos += 1) { // goes from 0 degrees to 180 degrees
         // in steps of 1 degree
         myservo.write(pos); // tell servo to go to position in variable 'pos'
-        delay(15); // waits 15ms for the servo to reach the position
+        delay(20); // waits 20ms for the servo to reach the position
     }
     
     for (pos = 180; pos >= 0; pos -= 1) { // goes from 180 degrees to 0 degrees
         myservo.write(pos); // tell servo to go to position in variable 'pos'
-        delay(15); // waits 15ms for the servo to reach the position
+        delay(20); // waits 20ms for the servo to reach the position
     }
 }