--- a/libraries/Servo/library.properties
+++ b/libraries/Servo/library.properties
@@ -1,0 +1,9 @@
+name=Servo
+version=1.0
+author=Infineon Technologies AG
+maintainer=Infineon Technologies AG <www.infineon.com>
+sentence=Servo library for XMC microcontroller series
+paragraph=This library allows you to control servos with a XMC microcontroller
+category=Device Control
+url=http://www.arduino.cc/en/Reference/Servo
+architectures=xmc,arm