This was completely written from scratch, with exception of the example
files, which are from Arduino's repository with some changes.

This is an essential library, as some default examples from the IDE use
the Servo library.

Also, it uses xmc's full potential, generating the PWM by hardware.