@techpaul 2020-07-07
There is at lest one issue in examples
~~~
const uint16_t maxAnalog = (1 << RESOLUTION) - 1; //Uses bitshift for efficient compile-time calculation
~~~

Whilst this is true for Write resolution it is not for Read resolution, as they are different. Read has max of 12 bit default 10 bit

wiringAnalog has additions to get this value in functions denoted on anlog section of wik]( https://github.com/Infineon/XMC-for-Arduino/wiki/Analog-Functions-and-Additions)

Namely
1. getAnalogReadMaximum
2. getAnalogWriteMaximum

Specifically added to help other functions that do PWM on percentage to scale properly