@danielstuart14 2020-07-07
Also, the delay in the loop is so the servo has enough time to change its position before receiving another command.