@danielstuart14 2020-07-07
Also, the delay in the loop is so the servo has enough time to get there, before receiving another command.
The sweep example wouldn't work, for example, if it didn't had a delay.