@danielstuart14 2020-07-07
About the issue with timing, it's likely that it was because of the wrong crystal frequency that was defined. It should be correct now.