@techpaul 2020-07-07
> Also, the delay in the loop is so the servo has enough time to get there, before receiving another command.
> The sweep example wouldn't work, for example, if it didn't had a delay.

Wrong way round the FASTEST RC style Servos should be updated (amongst common specs) is 50Hz i.e. 20ms. This needs be set in the setAnalogWriteFrequency when you attach the servo, and is then FIXED (unlike in AVR)

The AVR style with their overheads of chaining delays on timers without using standard channel timings means like quite a lot of  AVR style libraries it meets the specs under **SOME** conditions. A delay of 15 ms in the examples is TOO FAST when using Hardware PWM means you have a creeping update cyctl where at some point you will get an update happening either during a high pulse or worst case missed or causes jitter.

There are MANY things wrong with the AVR and other standard versions like resolution down to 4 micro seconds, I have used PWM down to 0.5 micro seconds on XMC1100 and XMC1300.

As Hardware PWM repeats the same signal setting UNTIL changed updating at 15 ms intervals defeats the purpose of using PWM for a single channel. If you have
1. delay 100ms (update rate 10Hz) it shows PWM continuing and actually works at human adjustable rates
2. You never get missed or other issues on steps
3. If you sweep on 3 degree steps at 100 ms delay per loop time on a sweep on a SIMPLER sweep algoritm gives 6 seconds 0 to 180 degrees.

Been testing this on examples using 2 and 4 channel outpts.

I have not checked all the code but the AVR method does NOT use min/max settings in attach beyond saving them. They use a range of 544 to 2400 which should be your default extremes. I attach a few doc files as azip that I created for where I was up to so far

[readme.zip](https://github.com/Infineon/XMC-for-Arduino/files/4887406/readme.zip)