@techpaul 2020-07-07
> I usually like using "this->" even when not needed, as it makes easier to tell that something is an instance variable by just glancing at it. But I'm going to change it.

This method is ONLY appropriate when passing references/pointers around, by using a class the method automatically has this-> passed in and only on very old C++ compilers was it necessary to show it was explicitly this class method. Scope of not file, local, or global means match with class if present.

I beleive you are copying errors unknown to you) from Arduino's bad code into your version, which is why I was doing some extensive tests.