@techpaul 2020-07-07
> 
> 
> About the issue with timing, it's likely that it was because of the wrong crystal frequency that was defined. It should be correct now.

Timing issues appear in many places my tests on timing were for servo output channels where scaling was wrong and NON-linear  600 to 1200 range being changed to 700 to 2000 with a multiply of 2.

I have been testing so far on Boot Kits for XMC1100 and XMC1300, not touched XMC4700 or XMC11002GO or others yet