@techpaul 2020-07-07
> Actually, about the default ms values, idk what Arduino does, but the industry default values are between 1000 and 2000ms for a single channel PPM servo.

NO that is what a lot of people use to limit on the SERVO it is **NOT the limit of the interface**

> There are many sources for this, you can check in the web, but here is one with it very detailed: https://dronebotworkshop.com/servo-motors-with-arduino/
> 
> Also from wikipedia: https://en.wikipedia.org/wiki/Servo_control#/media/File:Sinais_controle_servomotor.JPG

Wrong way round the standard for radio control and the maximum width is to do with transmitters and receivers and the PPM of the individual servo channles. this has always been an absolute maximum of 0.5 ms to 2.5ms plus sync pulse in 20 ms, which ALWAYS made 8 channel radio conytrol systems difficult.

Many SERVOS that RECEIVE their servo channel pulse once every 20 ms at fastest have these ays limits of 1ms to 2 ms for ease of the hardware/software/and channel compression for using methods such as PCM and other radio links. I have dealt with many transmitters, receivers, servos, or servo interface devices that have ranges 0.6ms to 2.4 ms. That is why many radio control transmitters(controllrs) had adjustable end stops on their joysticks to match servos.

The adding of min and max to the attach method was to limit the signal timing to match the servo limits for an individual servo, the interface has a spec of 0.5 to 2.5ms ABSOLUTE maximum.

Attach having min_angle and max_angle is different to standard Arduino and can cause conflicts in timing to degree and vice versa translations, in some cases